move boundary related functions to dedicated file
